+++ app/login/page.tsx
@@ -1,4 +1,4 @@
-import { AuthModalLauncher } from '@/components/auth/auth-modal-launcher'
+import { AuthModalLauncher } from '@/components/page/auth/components/auth-modal-launcher'
 
 export default function LoginPage() {
   return <AuthModalLauncher mode="sign-in" />
