+++ components/page/repos/repositories-page.tsx
@@ -1,0 +1,5 @@
+import { IssuesWorkspace } from '@/components/page/repos/components/issues-workspace'
+
+export default function IssuesPage() {
+  return <IssuesWorkspace mode="repos" />
+}
